Issue #44: expose fractional edit quantities

+++ src/ProjectHopeLynden.Application/Inventory/InventoryEntryEditView.cs
@@ -5,7 +5,7 @@ public sealed record InventoryEntryEditView(
     string ItemName,
     int CategoryId,
     int LocationId,
-    int CurrentQuantity,
+    double CurrentQuantity,
     DateTime? BestByDate,
     bool IsCommodity,
     bool IsMenuItem);
